[chore] Extracting toview
